Feature/fetch

+++ back/src/api/routes/want.js
@@ -1,26 +1,26 @@
-import { Router } from "express";
-import WantService from "../../services/want.js";
+import { Router } from 'express';
+import WantService from '../../services/want.js';
 const router = Router();
 
 const WantServiceInstance = new WantService();
 
 router.get('/', async (req, res) => {
-    const artwork_id = req.artwork_id;
+  const artwork_id = req.artwork_id;
 
-    try{
-        const wants = await WantServiceInstance.getArtworkWant(artwork_id);
-        res.status(200).json(wants);
-    }
-    catch(err) {
-        res.status(404).json(err.toString());
-    }
-})
+  try {
+    const wants = await WantServiceInstance.getArtworkWant(artwork_id);
+    res.status(200).json(wants);
+  } catch (err) {
+    res.status(404).json(err.toString());
+  }
+});
 
 router.post('/', async (req, res) => {
     const artwork_id = req.artwork_id;
-    const { user_id } = req.query;
+    const { user_id } = req.session;
 
     try{
+        if(!user_id) return res.status(401).json('Nonauthorized');
         const want = await WantServiceInstance.postArtworkWant(artwork_id, user_id);
         res.status(201).json(want);
     }
@@ -31,9 +31,10 @@ router.post('/', async (req, res) => {
 
 router.delete('/', async (req, res) => {
     const artwork_id = req.artwork_id;
-    const { user_id } = req.query;
-
+    const { user_id } = req.session;
+    
     try{
+        if(!user_id) return res.status(401).json('Nonauthorized');
         const want = await WantServiceInstance.deleteArtworkWant(artwork_id, user_id);
         res.status(201).json(want);
     }
